AAssetManager logic added to be able to load a shader file in Android.

+++ VknConfig/include/VknCycle.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint> // For UINT32_MAX
 #include "VknConfig.hpp"
 #include "VknEngine.hpp"
 
@@ -38,7 +39,7 @@ namespace vkn
 
         // Params
         uint32_t MAX_FRAMES_IN_FLIGHT = 0;      // Should match the value used in createSyncObjects
-        uint32_t m_defaultTimeout = UINT32_MAX; // Or std::numeric_limits<uint32_t>::max() if <limits> is included
+        uint64_t m_defaultTimeout = UINT64_MAX; // For vkAcquireNextImageKHR, timeout is uint64_t
         int m_width{0};
         int m_height{0};
         VkExtent2D m_extent{0, 0};
